+++ package.json
@@ -17,6 +17,7 @@
     "@babel/runtime-corejs2": "^7.1.5",
     "@cryptoeconomicslab/chamber-core": "^1.1.0",
     "@cryptoeconomicslab/chamber-wallet": "^1.1.1",
+    "@cryptoeconomicslab/plasma-chamber": "^0.0.10",
     "babel-plugin-transform-react-jsx": "^6.24.1",
     "bip39": "^2.5.0",
     "eslint": "^5.8.0",
@@ -29,6 +30,8 @@
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-redux": "^5.0.7",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "redux": "^4.0.0",
     "redux-thunk": "^2.3.0",
     "semantic-ui-react": "^0.84.0",
